// projects/design-system/src/lib/svg/svg-names.enum.ts
export enum SvgNames {
  LOGO = 'otp-logo',
  LOGO_SINGLE = 'otp-single-color',

  HOME_16 = 'OTP-icon-16x16-home',
  HOME_24 = 'OTP-icon-24x24-home',
  PAYMENTS_24 = 'OTP-icon-24x24-switcher-left-right',
  ACCOUNTS_24 = 'OTP-icon-24x24-bank',
  CARDS_24 = 'OTP-icon-24x24-credit-card',
  LOANS_24 = 'OTP-icon-24x24-loan',
  DEPOSITS_24 = 'OTP-icon-24x24-piggy-bank',
  TELEPHONE_24 = 'OTP-icon-24x24-telephone',
  CLOSE_11 = 'OTP-icon-11x11-close-x',
  SIGN_OUT_24 = 'OTP-icon-24x24-sign-out',
  ENVELOPE_24 = 'OTP-icon-24x24-envelope',
  GEAR_24 = 'OTP-icon-24x24-gear',
  ALERT_TRIANGLE_16 = 'OTP-icon-16x16-alert-triangle',
  ALERT_TRIANGLE_24 = 'OTP-icon-24x24-alert-triangle',
  DELETE_16 = 'OTP-icon-16x16-delete',
  HAMBURGER_24 = 'OTP-icon-24x24-hamburger',

  ARROW_DOWN_16 = 'OTP-icon-16x16-arrow-down',
  ARROW_UP_16 = 'OTP-icon-16x16-arrow-up',
  ARROW_LEFT_16 = 'OTP-icon-16x16-arrow-left',
  ARROW_RIGHT_16 = 'OTP-icon-16x16-arrow-right',
  ARROW_LEFT_24 = 'OTP-icon-24x24-arrow-left',
  ARROW_RIGHT_24 = 'OTP-icon-24x24-arrow-right',
  TRIANGLE_ARROW_DOWN_16 = 'OTP-icon-16x16-triangle-arrow-down',

  DOT_EMPTY_GREY_8 = 'OTP-icon-8x8-dot-empty-grey',
  DOT_FULL_GREEN_12 = 'OTP-icon-12x12-dot-full-green',

  SUN_24 = 'OTP-icon-24x24-sun',
  CHECK_CIRCLE_24 = 'OTP-icon-24x24-check-in-circle',
  ERROR_24 = 'OTP-icon-24x24-error',
  INFO_16 = 'OTP-icon-16x16',
  INFO_24 = 'OTP-icon-24x24-info',
  CHECK_CIRCLE_16 = 'OTP-icon-16x16-check-in-circle',
  REFRESH_16 = 'OTP-icon-16x16-refresh',
  MORE_ACTIONS_16 = 'OTP-icon-16x16-more-actions',
  HEART_16 = 'OTP-icon-16x16-heart',
  CLOSE_16 = 'OTP-icon-16x16-close',
}